requirements.txt
sentence-transformers==2.6.1
transformers==4.38.2
torch==2.2.2
faiss-cpu==1.8.0
numpy<2